Add candid interface (#16)
* Splits the `cargo-build` into its own separate job. This allows us to cache the rust targets, which will speed the e2e test workflows that will be introduced in future PRs.
* Adds rustfmt

+++ src/types.rs
@@ -13,7 +13,7 @@ use std::convert::TryInto;
 /// The payload used to initialize the canister.
 #[derive(CandidType, Deserialize)]
 pub struct InitPayload {
-    pub stability_threshold: u32,
+    pub stability_threshold: u128,
     pub network: Network,
 }
 
@@ -60,8 +60,11 @@ impl From<&BitcoinTxOut> for TxOut {
 
 #[derive(CandidType, Clone, Copy, Debug, Eq, PartialEq, Hash, Serialize, Deserialize)]
 pub enum Network {
+    #[serde(rename = "mainnet")]
     Mainnet,
+    #[serde(rename = "testnet")]
     Testnet,
+    #[serde(rename = "regtest")]
     Regtest,
 }
